Add close editor icon to api home screen and query home screen
