+++ app/client/src/pages/Editor/QueryEditor/QueryHomeScreen.tsx
@@ -17,10 +17,10 @@ import AddDatasourceSecurely from "./AddDatasourceSecurely";
 import { QueryAction } from "entities/Action";
 import CenteredWrapper from "components/designSystems/appsmith/CenteredWrapper";
 import DatasourceCard from "./DatasourceCard";
+import CloseEditor from "components/editorComponents/CloseEditor";
 
 const QueryHomePage = styled.div`
   padding: 20px;
-  padding-top: 30px;
   overflow: auto;
   display: flex;
   flex-direction: column;
@@ -29,6 +29,7 @@ const QueryHomePage = styled.div`
   .sectionHeader {
     font-weight: ${(props) => props.theme.fontWeights[2]};
     font-size: ${(props) => props.theme.fontSizes[4]}px;
+    margin-top: 10px;
   }
 `;
 
@@ -122,6 +123,7 @@ class QueryHomeScreen extends React.Component<QueryHomeScreenProps> {
 
     return (
       <QueryHomePage>
+        <CloseEditor />
         <p className="sectionHeader">
           Select a datasource to query or create a new one
         </p>
